Refactored old implementation
