+++ source/problemset-10/CMakeLists.txt
@@ -1,13 +1,13 @@
 cmake_minimum_required(VERSION 3.12)
 
 # Define all homework files
-set(TASKS_SRCS  task-10-graph-of-roads.cpp
-                graph.cpp
-                dataInput.cpp
-                dataOutput.cpp)
+set(TASKS_SRCS  Graph/task-10-graph-of-roads.cpp
+                Graph/graph.cpp
+                Graph/dataInput.cpp
+                Graph/dataOutput.cpp)
 
 # Define all tests
-set(TESTS unitTests/utest-task-10.cpp)
+set(TESTS   unitTests/utest-task-10.cpp)
 
 add_executable(problemset-10    ${TASKS_SRCS}
                                 ${TESTS})
