--- source/problemset-10/DataInput.txt
@@ -1,15 +1,0 @@
-9
-11
-2 7 3
-2 6 2
-7 8 7
-8 1 3
-8 9 1
-1 9 4
-5 9 20
-2 5 1
-4 6 7
-3 4 12
-3 5 15
-3
-1 2 3
