+++ source/problemset-10/note.md
@@ -4,14 +4,21 @@
         Есть множество городов и дороги, связывающие эти города.
         Для каждой дороги задана её длина.
 
-        Задача – распределить города между государствами по такому алгоритму:
-            задаются k столиц каждого государства,
+        Задача -> распределить города между государствами по такому алгоритму:
+            Задаются k столиц каждого государства,
             далее по очереди каждому государству добавляется ближайший незанятый город,
             непосредственно связанный дорогой с каким-либо городом,
             уже принадлежащим государству (столицей или каким-либо городом, добавленным на одном из предыдущих шагов).
             Процесс продолжается до тех пор, пока все города не будут распределены.
-            Граф дорог связный. Во входном файле: n – число городов и m – число дорог.
-            Далее следуют сами дороги в формате: i j len, i и j – номера городов, len – длина дороги.
+            Граф дорог связный.
+
+        Во входном файле:
+            n – число городов и m – число дорог.
+            Далее следуют сами дороги в формате:
+                i j len,
+                i и j – номера городов,
+                len – длина дороги.
+
             Далее задано число k – число столиц, далее – k чисел – номера столиц.
 
         Надо вывести на консоль номера государств и списки городов, принадлежащих государствам.
