+++ source/problemset-10/unitTests/utest-task-10.cpp
@@ -1,16 +1,15 @@
 #include <gtest/gtest.h> // Google Test Framework
-#include "../task-10-graph-of-roads.h"
+#include "../Graph/task-10-graph-of-roads.h"
 
 TEST(problemset10, task10)
 {
 	SUCCEED();
 }
 
-
 TEST(problemset10, task10_1)
 {
 	DataInput *di = readDataInputFromFile("DataInput.txt");
-	DataOutput *dOut = new DataOutput();
+	auto *dOut = new DataOutput();
 	buildWorld(di, dOut);
 	
 	FAIL();
